Move enter away from a layer-tap
This caused chaos with trying to enter numbers in a sensible way.

Enter has moved to the inner right thumb key.
The middle right thumb key is now only a momentary layer change on
what was the lt number layer key.

They layout was also generated with https://nickcoutsos.github.io/keymap-editor/
which is the source of much whitespace change.

Esc is now a combo key to make things a bit simpler.

* Added BSPC to numbers layer
* Added < and > to code symbols layer to allow -> to be easier
  to type
* SPACE everywhere is now space / LGUI