Put a sticky shift on right thumb home
This adds &smo as a sticky-key-momentary-layer, which is used to
change the raw mo 2 into a universal one-shot shift key in addition
to mo 2. This gets around the issue with the shift HRM that keeps
being missed when typing. The next change is probably to make shift
modify a tap to produce &caps_word.