Move combos off the home row
tab and bspc_del combos both move from the home row to the bottom row.

This prevents them from fighting with the HRMs and requiring a long delay
before Ctrl-Shift can be applied.

Note that the num_word doesn't play nicely with combo layers - if the
layer is sticky, pressing another layer key keeps it up, so I fire off
<F7> when I try and send left-arrow while the sticky num layer is active.

Either the function layer needs to not be conditional, or canceling
becomes second nature.