Restore tab and bspc combos to home row
Applying more learnings from the urob zmk-config repo - make the combo
that sits on top of a hrm do the hrm hold that the keys making it up
would do. This means that the tab combo on the left hand sends
`lctrl+shift` when held and tab when tapped. This doesn't require any
waiting for mods to combine, and fires instantly.

This was tested with a vscode developer mode debug keyboard shortcuts
that shows the keypresses and decisions it makes about the key presses
it receives.

The tab just uses a hml for the combo behaviour, the bspc needs a
bespoke hmr behaviour to swap the tap &kp for the &bspc_del mod_morph
behaviour.