Add a commented out debug logging configuration
This was used to try and figure out why &caps_word sometimes fails to be
registered as a shifted letter in X11. Having tried a few things,
caps_word seems to work reliably on iOS, Windows, Linux console
terminal, VSCode in Linux, but not other X11 apps (terminal, alacritty,
text editor). This happens over USB and Bluetooth, and when USB logging
is enabled.

Capturing logs of the letter a being interpreted as "a" and "A" showed
no difference in the ZMK logs, but it seems more likely to occur on HRM
keys than kp keys, and seemed to stop on USB when the USB logging was
playing catchup, so probably a timing issue on the X end reading the
events out of order.