Swap ESC and bspc_del combos
This moves backspace off the home row, which makes it behave more
simply - no longer combining HRMs with combo keys. ESC is rarely
held down and is a plain key, whereas bspc_del on HRM is a mod-tap
combo / mod-morph beastie.

The next step may be to move tab also off the HRM.