Move Colemak layers into Win and Mac layer groups
This creates win and mac colemak layers that sit above the base layer
and share the symbol, nav and function layers of each group.

Start in Qwerty, use Function layer to select to other group's qwerty
or between this group's colemak and qwerty layers.

This keeps Undo/Cut/copy/paste keys on the NAV layer in the Qwerty
positions for C-Z, C-X, C-C, C-V.